+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 
 - Implemented Babel view, to compare translated items in add and edit mode. @giuliaghisini
 
+- Improve the blocks engine by adding a detector for clicking outside in the `BlocksForm` @sneridagh
 - Include a pluggable architecture for pluggable render-time insertions (similar to <Portal>) @tiberiuichim
+
 ### Bugfix
 
 - Include selected block in multiselections @sneridagh